Keep agent security tests inside semantic boundaries
Move the Worker-backed zero-length Unicode OutputFilter regression out of the colocated unit file so the semantic unit-boundary audit no longer classifies validator.test.ts as an undisposed effect-bearing unit. Keep the timeout harness unchanged in integration coverage and make the remaining test doubles explicitly return promises instead of relying on async callbacks with no awaits.

Constraint: The semantic audit forbids growing scripts/test/test-semantic-audit-migration.ts for new unit-boundary effects.

Rejected: Add validator.test.ts to the semantic migration inventory | the CI failure explicitly requires moving the effect or making the unit hermetic.

Confidence: high

Scope-risk: narrow

Tested: deno fmt --check src/agent/middleware/security/validator.test.ts tests/integration/security/output-filter-worker.test.ts

Tested: deno lint --config=scripts/test.deno.json src/agent/middleware/security/validator.test.ts tests/integration/security/output-filter-worker.test.ts

Tested: deno task lint:test-semantic-dispositions

Tested: deno task test:file src/agent/middleware/security/validator.test.ts tests/integration/security/output-filter-worker.test.ts

Tested: deno task lint:ci

Tested: git diff --check

Not-tested: Full deno task test after the lint-only CI fix.
